only restart the broken ice connection, not both
